New Change 22:14

+++ include/task1.h
@@ -4,6 +4,6 @@
 
 #ifndef A_TASK1_H
 #define A_TASK1_H
-unsigned int GetPrime(unsigned int n);
-bool IsPrime(long int a);
+int GetPrime(int n);
+bool IsPrime(int a);
 #endif //A_TASK1_H
