+++ src/task1.cpp
@@ -1,6 +1,6 @@
 #include "task1.h"
 
-unsigned int GetPrime(unsigned int n) {
+int GetPrime(int n) {
     int count = 0;
     int i = 2;
     while(count != n) {
@@ -12,8 +12,8 @@ unsigned int GetPrime(unsigned int n) {
     return i-1;
 }
 
-bool IsPrime(long int a) {
-    for(long int i = 2; i < a; i++) {
+bool IsPrime(int a) {
+    for(int i = 2; i < a; i++) {
         if(a % i == 0) {
             return false;
         }
